requirements.txt
streamlit
pandas
scikit-learn
pillow
pytesseract